fix repo URI in cargo.toml. move code from ocsp.rs to lib.rs.
