+++ x509-ocsp/Cargo.toml
@@ -7,7 +7,7 @@ Online Certificate Status Protocol - OCSP formats as described in RFC 6960
 """
 authors = ["RustCrypto Developers"]
 license = "Apache-2.0 OR MIT"
-repository = "https://github.com/RustCrypto/formats/tree/master/x509"
+repository = "https://github.com/RustCrypto/formats/tree/master/x509-ocsp"
 categories = ["cryptography", "data-structures", "encoding", "no-std"]
 keywords = ["crypto", "x.509"]
 readme = "README.md"
